+++ R/date2week.R
@@ -6,7 +6,8 @@
 #' @param week_start a number indicating the start of the week based on the ISO
 #'   8601 standard from 1 to 7 where 1 = Monday OR an abbreviation of the
 #'   weekdate in an English or current locale. _Note: using a non-English locale
-#'   may render your code non-portable._
+#'   may render your code non-portable._ Defaults to the value of 
+#'   [get_week_start()]
 #' 
 #' @param floor_day when `TRUE`, the days will be set to the start of the week.
 #'
@@ -45,7 +46,7 @@
 #'
 #' @author Zhian N. Kamvar
 #' @export
-#' @seealso [as.Date.aweek()], [print.aweek()]
+#' @seealso [set_week_start()], [as.Date.aweek()], [print.aweek()]
 #' @examples
 #'
 #' ## Dates to weeks -----------------------------------------------------------
@@ -57,6 +58,12 @@
 #' 
 #' # By default, the weeks are defined as ISO weeks, which start on Monday
 #' print(iso_dat <- date2week(dat))
+#' 
+#' # This can be changed by setting the global default with set_week_start()
+#' 
+#' set_week_start("Sunday")
+#'
+#' date2week(dat)
 #'
 #' # If you want lubridate-style numeric-only weeks, you need look no further
 #' # than the "numeric" argument
@@ -69,13 +76,21 @@
 #' # `floor_day = TRUE, factor = TRUE`:
 #' date2week(dat[c(1, 14)], floor_day = TRUE, factor = TRUE)
 #'
+#'
 #' ## Weeks to dates -----------------------------------------------------------
 #'
 #' # The aweek class can be converted back to a date with `as.Date()`
 #' as.Date(iso_dat)
 #' 
-#' # If you don't have an aweek class, you can use week2date():
-#' week2date("2019-W01-1", week_start = "Monday")
+#' # If you don't have an aweek class, you can use week2date(). Note that the
+#' # week_start variable is set by the "aweek.week_start" option, which we will
+#' # set to Monday:
+#' 
+#' set_week_start("Monday")
+#' week2date("2019-W01-1") # 2018-12-31
+#'
+#' # This can be overidden by the week_start argument;
+#' week2date("2019-W01-1", week_start = "Sunday") # 2018-12-30
 #'
 #' # If you want to convert to the first day of the week, you can use the 
 #' # `floor_day` argument
@@ -109,7 +124,7 @@
 #' # Epiweek definition: Sunday -- 7 
 #' date2week(dat, 7)
 #' date2week(dat, "Sunday")
-date2week <- function(x, week_start = 1, floor_day = FALSE, numeric = FALSE, factor = FALSE, ...) {
+date2week <- function(x, week_start = get_week_start(), floor_day = FALSE, numeric = FALSE, factor = FALSE, ...) {
 
   format_exists <- !is.null(list(...)$format)
   nas <- is.na(x)